su-tunnel for NetBeans added.
This is an extended version of Noel's original script, which was made
since I could not find a way to run the script with the currently opened
file. What it does? It checks which .rb script (in specified folders)
was changed last and it then runs it.

If someone has a better way, please let me know!