Fixed sketchup_build and updated README.
I needed to fix calling of tunnel_ide, as arguments were not passed
correctly. I achieved this by specifying ARGV, which namely throws a
warning, but it works correctly. If someone knows how to surpress this
warning, please fix it or let me know how to do it.

I also updated readme with information on how to set su-tunnel in
Netbeans. If you miss some information, please let me know and I will
update it.